Sépration des classes, En attente necessite node
