+++ src/Procedure.js
@@ -1,0 +1,15 @@
+class Procedure extends Probleme {
+    // ATTRIBUTS -non-
+    // CONSTRUCTEUR
+    constructor(abscisse, ordonnee, libelle, listeDonnes = [], listeResultats = []) {
+        super(abscisse, ordonnee, libelle, listeDonnes, listeResultats);
+    }
+
+    // ENCAPSULATION -non-
+
+    // METHODES
+    afficher() {
+        super.afficher();
+    }
+} window.customElements.define("procedure-element", Procedure);
+export default Procedure;
