+++ src/Test.js
@@ -1,0 +1,8 @@
+export default class Test
+{
+    constructor(name, age)
+    {
+        this.name = name;
+        this.age = age;
+    }
+}
