+++ src/index.html
@@ -11,7 +11,7 @@
 
     </plan-travail>
 </body>
-<script src="main.js"></script>
+<script type ="module" src="main.js"></script>
 <script>
     // Set css variables
     document.body.style.setProperty('--sizeModifier', 1);
